Add morphology tests
